Disable packing input/output for dynamic component index
Commit 8cd0205 caused a regression in tessellation in the Sascha WIllems
tesselation example. The root cause is not taking dynamic component
index into consideration. This change is a wordround for this case by
totaly disabling input/output packing. A proper fix should be a partial
packing scheme that is non-packable locations (with dynamic location
offset or component index) should be skipped and placed at the back of
packable locations with consecitve location index.
Considering a future required change of refactoring the location maps, I
think it' better to integrate the partial packing in that change.