Reapply: Add Image resource type workaround
GFX10 hardware cannot handle invalid image resource type [1, 7], so we
change such kind of value to 0 (buffer) to let hardware ignore the
request.  This may happen when buggy application declares an image in
shader but provide a buffer descriptor in driver. Also add a pipeline
option to disable it for possible performance loss.

This fix was originally by Ruiling Song as #909, but had to be
temporarily reverted as it showed up an unrelated problem in LLVM
waterfall loop support. It can now be reapplied.

I have made a couple of changes to Ruiling's fix:

1. Added a workaround flag, which is set only for GFX10.1. The fix is
   unnecessary for GFX10.3.
2. Saved one instruction by spotting a type field (bits 31..28 of dword
   3) with value [0,7] by just checking if dword3 is non-negative.

Change-Id: I13875268d03bd4ae41414bcee729af087db3ff47