Add new targets for a few cases in old hardware
Ready for a forthcoming PAL change that will start using them, this
adds:

gfx6.0.2: This separates Hainan and Oland out of gfx6.0.1. Unlike
          gfx6.0.1, the new target does not need the shaderZExport
          workaround.

gfx7.0.5: This separates a variant out of gfx7.0.3 that does not need
          the shaderSpiCsRegAllocFragmentation workaround.

gfx8.0.5: This separates TongaPro, the workstation variants of Tonga,
          out of gfx8.0.2. That does not make any difference to LGC; the
          LLVM AMDGPU back-end could in theory use it to know that the
          target has fast 64-bit shift instructions.

gfx9.0.12: Separates a variant out of gfx9.0.9.

Requires LLVM patch https://reviews.llvm.org/D90419 to be able to use
gfx9.0.12.

Change-Id: Ia3f2abf25a1d2048fec746b15c7acd64adf8d70c