Refactor the update of input/output location maps
The current update of location maps (i.e, inputLocInfoMap,
outputLocInfoMap, perPatchOutputLocmap and perPatchInputLocMap) includes
five steps:
- Inititalize the maps with InvalideValue in
InOutBuilder::markGenericInputOutputUsage
- Record the active locations and dynamic indexing for subsequently
removal use in PatchResourceCollect::visitCallInst
- Remove inactive input locations from inputLocInfoMap and
perPatchInputLocMap in PatchResourceCollect::clearInactiveInput
- Remove inactive output locations from outputLocInfoMap and
perPatchOutputLocMap in PatchResourceCollect::matchGenericeInOut
- Update the value of the four maps with continuous locations at the end
of PatchResourceCollect::matchGenericeInOut
Actually, the five steps are not required for the packable input/output
maps because they are re-built in packInOutLocation. Thus, this change
aims to integrete the four of the updating steps in one place only for
nonpackable input/output use. This change includes:
- The state of packing or unpacking the input/output of each stage is
recorded in m_inOutPackStates[][2]. We check it to perform packing or
unpacking on the input/output.
- For unpacking case, we add a new function
PatchResourceCollect::updateInOutLocMap to complete the functionality of
updating the maps. The main idea is to remove unused location from the
maps and then map the locations to be continous locations. For the
special locations that are using dynamic index or only used for output
import, they are put at the back with continious indices. The code of
above metioned record/remove/update is reomved. The data of used input
import calls and output export calls and output import locations are
collected in PatchResourceCollect::visitCallInst.