Disable CS fixed user data layout restriction
PAL used to have restrictions on user data layout in a compute shader
that have now been lifted (as of PAL version 624). This commit takes
advantage of that in LGC, giving the same freedom in user data layout as
in a graphics shader.

That will be required when shader compilation is extended to compute
shaders, as the user data will not be available at compile time to apply
the restrictions.

Change-Id: I0432e35de9c255bb5bd77014bfbc92203aad4adf