Remove all instructions in a block after OpKill.
During inlining it is possible to construct a block which has
trailing instructions after a kill.  This can prevent return
block unification and cause exports to be skipped.
In particular some SPIRV generators terminate a block with a
function calling OpKill with OpUnreachable, this also fixes
these by replacing unreachable with ret void, which is then
unified.