Account for relocs when improving index choice for waterfall loops
Introduction of relocs meant that the code to improve waterfall loop indices no
longer worked. This meant it would always fall back to the value tagged as
nonUniform (usually a resource descriptor).

The pre-existing code to attempt to improve the index to the offset used in the
GEP for the resource descriptor load (it is much smaller) broke because it could
trace back through a reloc and associated code to manipulate it.