Use the original key when removing unused outputs
Commit 966a82736b0df541c65108d40905e707b553c139 merged GsOutLocInfo into
InOutLocationInfo and made 'location' start at bit 3 instead of bit 0,
exposing a bug where the GS output location would be used as the map key
when removing unused outputs. Use the original map key instead of the raw
location when removing the unused outputs.

For example, if a GS has 12 outputs but only 0-4 are used then when
output location 8 (b1000) is removed it will actually remove output
location 1 when interpreted as an InOutLocationInfo.