Consider waterfall.last.use intrinsic in image rsrc workaround
In the case where an image resource descriptor is required to have a workaround
applied, we additionally need to consider the case where a waterfall.last.use
intrinsic is used, otherwise, the resource descriptor will be adjusted and it
will fall outside the waterfall loop (the last use becomes the first use in the
workaround code and not the image intrinsic).

We now additionally process last.use to see if that is for an image resource
descriptor, apply the workaround to the last.use. This means that the intent of
the waterfall code is preserved (the last use is the image intrinsic still).