Refactor vertex attribute export
We didn't collect vertex attribute exports previously. This has a latent
issue (see below GLSL of a VS):

  layout(location = 0, component = 0) out vec2 data1;
  layout(location = 0, component = 2) out vec2 data2;

We finally generate two exp instructions with this:

  exp param0, data1.x, data1.y, off, off
  exp param0, off, off, data2.x, data2.y

So this change is to collect vertex attribute exports and do exporting
with gathered operations. We introduce two functions:

  recordVertexAttribExport, exportVertexAttribs

Their functionalities are apparent. We don't explain more here. This
refactoring work is also for a future use where vertex attributes are
handled differently.

Change-Id: Ibed717a06ffd0b04ab3eaf5332e83148e795ee1d