Support partial packing with dynamic indexing on VS-TCS
TCS input import calls with dynamic indexing (locOffset/elemIdx) and the
corresponding VS output export calls cannot be unpacked - call them
unpackable calls. We should keep unpackable calls unchanged and map
their location properly. There are three things to do:
1.Skip unpackable calls in scalarization input calls
2.Skip unpackable calls in re-asssemble output calls
3.The inputLocInfoMap/outputLocInfoMap use plain location for unpackable
calls; The locations of the packable calls are remapped in front of
those of the unpackable calls.
4.The key to look up VS outputLocInfoMap or TCS inputLocInfoMap can be
the combination of location and component or just location.