Support new Vulkan color attachment formats
Those new formats are not added to color attachment mapping:
  VK_FORMAT_A4R4G4B4_UNORM_PACK16_EXT
  VK_FORMAT_A4B4G4R4_UNORM_PACK16_EXT

They are mapped to 4_4_4_4 data format plus UNORM numeric format.

Our mapping table is unable to support those formats that belong to
Vulkan extensions since the token value is too large as the table
array index. This, use a std::unordered_map to achieve this.

Change-Id: Ic733ff3b976c1d6ff225376c71d060355bb839a2