Elf linking: use float mode from shader body
When merging the pal metadata, we currently do an xor of the float
modes. This can cause problems when the fetch or color export shader
have a different float mode than the body of the shader. The solution
is to use the float mode from the body of the function.

Fixes https://github.com/GPUOpen-Drivers/llpc/issues/1047.