Allow creating exp instruction for tranform feedback output in copyshader
In current visitReturnInst of PathInOutImportExport pass, tranform feedback output of copyshader will not create "exp" instruction for generic output. This condition is added two years ago to fix a dxvk issue to support xfb extension. We want to remove the restriction to verify if the condition is really required. If it is required, we will make a change to avoid useless "lgc.output.export.generic" generated in PatchCopyShader pass for xfb output.