Pack input/output on GS pipeline
Support packing on the pipeline that includes geometry shader. The
mainly changes are to implement the packing:
-{VS/TES} output -> GS input: use GS input to build the location info
map to fill GS outputLocInfoMap and copy to the previsou stage
inputLocInfoMap. This process is similar to VS-TCS.
-Gs output ->Copy shader output: this is a new process due to transform
feedback export. First, scalarize xfb outputs to build the
`locInfoXfbOutInfoMap` in InOutBuilder to corresponding to the key of GS
outputLocInfoMap which is built by scalarized generic output exports.
Secondly, collect the byte sizes of the mapped location to load output
value from LDS at the mapped location. Lastly, the obtained output value
at the mapped location is the packed result, so we can create export
generice call directly; For xfb export, we need an extract-insert
process to re-build the original xfb output value with respect to the
xfbOutInfo.