Allow exp instruction for xfb output of copyshader
In current PathInOutImportExport pass, "exp" instruction will not be
generated in copyshader for generic outputs if transform feedback is
enabled. This restriction is enforced to fix a dxvk issue in early year.
But the fragment shader is allowed its input receiving xfb output of the
copyshader. Remove this restriction in this change.