Amend PatchLoopMetadata pass
Several changes to the loop metadata:

Replaced the -strict-unroll-hints and -strict-dont-unroll-hints boolean options
with -unroll-hint-threshold and -dontunroll-hint-threshold respectively.
The new options take an integer threshold value. A value of 0 (default) will use
the llvm.loop.unroll.full or llvm.loop.unroll.disable metadata as previously.
When non-zero the amdgpu.loop.unroll.threshold metadata with the specified value
is attached to loops wih the SPIR-V "Unroll" and "DontUnroll" hints.

Added option -disable-licm-threshold which takes an integer value - a non-zero
value will disable the LICM pass for any loops with at least the specified number
of blocks. The -disable-licm option is retained for backwards compatibility, but
is now equivalent to -disable-licm-threshold=1.
A default value of 20 is used - which disables LICM for loops with 20 blocks or
more. This allows LICM to operate in most cases where it is beneficial, but
significantly reduces register pressure resulting from LICM on larger loops (and
thereby increases occupancy).