Fix a latent issue in copy shader when multi-stream is enabled for XFB
In the pass of in/out import/export, we always assume the output export
happens in the return block. However, it is not sure for copy shader.
When multi-stream is enabled for XFB, the export stays in corresponding
handling branches for this stream. Thus, all processing in the return
block for in/out should stay in the same place.

Change-Id: I831a1b5a24dc6a63a76435ec031738ecee2a9551