Separately report cache hit of internal cache
`Compiler::BuildGraphicsPipeline/BuildComputePipeline()` reports cache
access results to the [Pipeline Creation Feedback](
https://www.khronos.org/registry/vulkan/specs/1.1-extensions/html/chap11.html#pipelines-creation-feedback).
Since the pipeline cache and base pipeline acceleration are only two
options to report the cache access result (see
https://www.khronos.org/registry/vulkan/specs/1.2-extensions/man/html/VkPipelineCreationFeedbackFlagBitsEXT.html),
we have to separately report the cache hit of the internal cache.
This commit separates the cache hit of the internal cache from the
cache hit of the pipeline cache that is given by VkPipelineCache.