Favour uniform uses
Use the uniform (non-divergent) node instead of a divergent one
if they are equal.

Detect a case where a branch condition tests two nodes for equality
where one of them is divergent and the other uniform. In the true
block replace the divergent use with the uniform one as it helps
instruction selection decide what to put in an SGPR.