Fix buffer descriptor indexing with NonUniform
The argument `isNonUniform` of `CreateLoadBufferDesc` is determined by
checking the users of the block index for any NonUniform calls. If has
NonUniform call, the flag is true, otherwise is false. This way will
give a wrong answer for the case that `NonUniform` is decorated on a GEP
instruction not the index in SPIR-V shader. To fix it, we need check the
users of GEP itself for any NonUniform calls before checking the users
of the index to avoid creating readfirstlane on the index.

The SPIR-V shader can be written as below:
OpDecorate %21 NonUniform
...
%17 = OpLoad %int %rIndex
%18 = OpCopyObject %int %17
%21 = OpAccessChain %_ptr_StorageBuffer_v4float %data %18 %int_1