Use patch control point only when GS/TCS/TES is given
Based on my code analysis, the patch control point may have an impact on
the geometry shader, tessellation control shader, and tessellation
evaluation shader ELFs. However, it does not have any impact on the
ELF of a vertex, compute, or fragment shader. Therefore, we have to
excludes the patch control point data from the hash.

The initial motivation of this fix is to make the hash value of amdllpc
and LLPC used for ICD the same. The existing
`PipelineDumper::updateHashForNonFragmentState()` includes
patch control point data to generate the hash for non-fragment shaders.
Since amdllpc sets `patchControlPoints` to 3 when its given value is 0,
amdllpc has a different hash value from LLPC used for ICD. This CL
fixes the issue.