LGC: Remove NGG fast launch, and simplify LGC primitive type
NGG fast launch is deprecated, so remove it.

That then allows a slight simplification of the LGC interface. The only
reason LGC needed to know the primitive topology (triangle strip,
triangle list, etc) rather than just the primitive type (triangle etc)
was that handling NGG fast launch needed to know the number of vertices
per primitive.

So this commit changes the LGC interface to pass only the primitive
type.

Change-Id: I21666bb08c92590e6757ea34df7a2e80f79f0d6a