Add .clang-tidy configuration
This file is picked up by clang-tidy, which is used e.g. by the clangd
language server. When used with an IDE, it will show hints when the
naming scheme is not followed (e.g. member variable not starting with
m_) and even offer automated fixes.

The content of the file is a merge between the LLVM .clang-tidy and
scripts/switch-coding-style.sh.

I always forget the coding style after working on other projects, this
should automatically remind me.