Revert "Add cache hash note to shader ELF"
This reverts commit e6247f101f0f2dfc07054866a571673650c84655.

The required cache hash info is now written to and used from PAL
metadata and this option to write to the notes section is no longer
needed.

This is not a complete revert. Its new test (which has already been
updated), additional log statements used by tests, and other logic that
is still in use, have not been removed.