Set the user data limit is set to max when there is user data spill
When linking shader, the user data limit is set to the max only if no
user data nodes are used, but there are some.  This is insufficiant
because it needs to happen whenever the spill table is used.

This commit moves the code that set the user data limit when spill is
used to when the pipeline is finalized.  That way it will be called on
all paths.