Fix msgpack dumper for empty array
Previously it was crashing with my new LGC part-pipeline compilation
scheme, trying to compile
llpc/test/shaderdb/ObjInput_TestFsBuiltIn_lit.frag (which has no generic
inputs so generates an empty array) with the -v option.

I guess it might need an analogous fix for an empty map, but I didn't
want to go as far as that as I do not have a test case for it to hand.

Change-Id: Ib78b5f6fa3b3c109f6f990d89adc812834eaec9e