Add ELF disassembler in LGC
Could be used to dump ELF that LLPC generates, instead of using the
non-LLVM-based ELF dumper in the LLPC front-end source. Could also be
used to implement the "show me the ISA disassembly" Vulkan extension,
instead of relying on the AMDGPU option to embed disassembly in the ELF.

Also builds an lgcdis standalone tool that disassembles an ELF with the
ability to decode the PAL metadata.  The intention is that the output of
the disassembler is re-assemblable to create a valid PAL pipeline ELF.
However that has not been tested.

A further goal would be to generate more usable disassembly, with a
symbol, for a multi-instruction sequence starting with s_getpc for a
long branch or an access to constant data. Doing that would need
enhancements in the disassembler in the AMDGPU target, and possibly the
addition of a new "disassemble a whole blob of code" API to the
MCDisassembler interface.

Change-Id: Iaec3f4b62d6af58b96838eee39e69e62fade9e4e