Support OpSelect between block arrays in lowerBufferBlock
In current `lowerBufferBlock`, we only consider the users of the global is either a GEP or a bitCast when the global is an array of blocks. There are cases that the user can be a select instruction via variable pointers then the select is referenced by GEPs. This change will support this case. We should lower the true and false value of the select and propagate the GEP's block index and buffer flags to invoke `CreateLoadBufferDesc` for the two globals.