Remove redundant passes in lowering
Both DCE and SROA were run earlier in the same pass list and they
do not need to be run again.

This patch does not change the generated code when tested on a
large sample of real-world shaders.

Note: we probably do not need to run simplifycfg twice here either,
but earlycse opens up some more opportunities for cfg simplification,
so it is not that clear.