Remove redundant passes in patching
Remove passes that are not suited to our target or have
been run already so there is no use running them again.

This patch does not change the generated code when tested on a
large sample of real-world shaders.

This does not address the elephant in the room: instcombine and
also peephole, but rationalizing them changes generated code, so
it will be done separately.