Add interface to check the caches in LLPC.
The code to check the caches is repeated in three places.  This code is
also hard to read and easy to get wrong.  The cache handles have to be
expliticly released or it could mess up checking the cache at other
times.

The code is also complicated by the fact that there are potentially 4
different caches with two different types: The ICache and the
IShaderCache.  This can lead to inconsistencies in how the various
caches interact.

I'm proposing that we add a class that will make take care of all of
this, and hopefully make the code easier to read and less error prone.