Force to do per-smaple interpolation in some case
This change is a patch for interpolation.When per-sample shading is
enabled, force the nonperspective and smooth input with center-based
interpolation and explicit vertex parameter input to do per-sample interploation.
The patch is disabled by a temporary option `DisableInterpModePatch`.
When the Blending and MSAA tests is updated, the option will be removed
to enabled the patch.