Reduce the number of instcombine iterations
The patch reduces the compilation time by over 3% on average
and does not cause a perf regression in my tests.

My experiments have shown that having instcombine in several
places in the pass manager is good for code quality.
However, since we run it multiple times, we can decrease MaxIterations
limit on each run.