Refactor relocatable shader compilation
To match other work that is being done, we want relocatable shader
compilation to be divided into vertex processing and fragment stages.
The current implementation tries to compile each shader stage separatly.
This is currently a difference without distinction because only handle
CS and VsFs pipelines.

When we will try to handle geometry shader, we want to compile the vertex
shader and the geometry shader together.   We want to refactor the
relocatable shader code to make that easier.