Move PatchResourceCollect::m_inOutPackState to PipelineState
The change aims to ensure the generic input/output of a stage can go
througth the same handling either packing or no-packing in different
places. This will fix some potential issues.
In InOutBuilder::CreateWriteXfbOutput, the generic XFB output is
scalarized without considering the packable state while the generic
output of GS is scalarized based on the packable state, which will cause
mismatch.