Port the SpirvLower class and SpirvLowerTranslator pass to the new pass manager
This patch splits the SpirvLower class, which is the base class for all the
lower passes, into two separate classes: (i) LegacySpirvLower is used as the
base class for all the lower passes implementing the legacy pass manager
interface. (ii) SpirvLower is used as the base class for all the lower passes
implementing the new pass manager interface. This split is designed in a way
that once we fully switch to the new pass manager, we'll only need to remove
the Legacy* classes from the tree and no additional work will be required.
The patch also ports the SpirvLowerTranslator pass to the new pass manager
using that new SpirvLower class.