Change immutable samplers in pipeline state to plain data
LGC pipeline state was keeping an IR constant (an array of <4 x i32>) for each
immutable sampler. It is a bit dodgy keeping an IR value outside of IR
beyond a single pass, and we were running into problems.

This commit changes pipeline state to keep immutable samplers as plain
data. It gets converted to an IR constant in DescBuilder when it
constructs the constant global variable to represent the immutable
sampler.