Place the merge Es shader at the start of the text section.
When linking unlinked shader in LGC, the vertex fetch shader will be
prepended to the text section.  So the start of the text section must
contian the function to which the fetch shader should be added.  This is
why we want the merged Es shader at the start of text section.