Add the new pass manager class
This patch adds the (still WIP) new PassManager class. This class implements
LLVM's new pass manager interface. Unlike in the legacy PassManager class, we
inject custom code between passes using the new instrumentation callback system
provided by LLVM. We also rely on LLVM's standard instrumentation classes to
implement the -verify-ir option. There are currently two known issues with this
new implementation: 1. --dump-cfg-after does not work because there is no
mechanism to register a pass with an ID to the instrumentation system.
2. --time-passes does not work because we need to use the standard
instrumentation classes to make it work. These two issues will be addressed in
subsequent patches when we'll have more passes ported to the new PM to
experiment with. The patch also adds the -experimental-new-pm flag that is used
to test the new pass manager implementation. The option will eventually be
removed when the switch is complete. One of the uses of the pass manager in
llpcCompiler.cpp is ported to that new PassManager class and is guarded by that
new experimental flag.