Remove the resource collect pass from lowering
It seems that this pass is used to collect information about modules and to
remove dead global variables. In LegacySpirvLower::addPasses, the pass is almost
directly followed by LLVM's global DCE pass that can also remove these globals.
Also, it seems the collected information is never used for this instance, so the
pass can probably be removed from there. I tried compiling 10K+ shaders with and
without the pass and the generated code is exactly the same in all cases.