Remove output export call if the output value is undefined
This is to remove unnecessary output export calls. These calls are
generated when we lower aggregate-typed outputs. For example, if
part of a structure-typed output are written, we export the whole of
it but some members are dead outputs. There is no need of doing this
and the dead output export calls will waste HW instructions, such
as LDS/buffer write.

Change-Id: I3b07eab59c42f080dca12acd51e9fe2dafb20bd3