Move the out of bounds check to the end of the SPIRVReader to avoid various problems:
We found out that there were some issues with control flow changes which can occur in the SPIR-V code. For instance, the current OOB check doesn't work / produces invalid LLVM IR when an control flow change occurs (OpLabel, some instructions, OpBranch to some other label which gets inserted later, then an OpPhi node gets discovered) while in this first label a store or load instruction gets wrapped by the OOB guard. The problem here is that the translated Phi node has no knowledge of the actual control flow change due to the sequential translation logic and therefore will not have its operand correctly set to the actual predecessors which are changed according to the OOB logic:

```
%27 = OpLabel ; .entry
OpStore %34 %33
OpStore %35 %33

OpBranchConditional %36 %37 %38
%38 = OpLabel ; .some other BB
%39 = OpPhi %int %int_0 %27 %40 %41
```

The Phi returns 0 when the predecessor is the entry point and some other value otherwise.
However, inserting the bounds check would cut off the LLVM IR here, insert a conditional branch as terminator for the current block, move the OpStore to a conditional block and the rest of the code to a final block which will get reached unconditionally.
The OpPhi node relies on label %27 (entry block), it will not change its operands (we are still doing a direct translation from SPIRV here). Because the conditional branch was moved to another BB because of the OOB checks, everything gets messy. Operand 1 of the OpPhi node should have the new basic block where the OpBranchConditional resides in now as predecessor, but we cant refresh the Phi node because at the point of inserting the OOB check to the store we dont know there is a dependency like that.

Also, the current implementation does not handle OpMatrix and OpVector instructions as well as constructs like arr[arr[dyn_index]]. Also, sequences of instructions which get created during an addLoadInstRecursively / addStoreInstRecursively call should be moved as a wholesome.

So, the intention behind this change is to record all instructions which get created during execution of one of the OpLoad / OpStore translation methods, then store the original SPIR-V instruction according with some meta information (are we handling loads or stores)? and a vector of translated LLVM instructions and then insert the OOB checks at the end of the function translation.
The actual splitting logic is done by SplitBlockAndInsertIfThen, which also takes care of updating Phi nodes appropriately by using the dominator tree. To correct dependent load results (e. g. we want to use a phi node as input for another load after translating a load instead of just doing a dumb split again for the other load), this is updating the m_valueMap for cases where this is relevant.

This also adds some more tests.