Don't update output loc info for undef outputs.
When there is an explict store to a vertex shader output that is an
undef value, then it will be removed.  That is fair.  However, the
output location map is updated to remove this output dropping the
channel that is used for other output.  This does not work for unlinked
shaders because the pixel shader will then read its inputs from the
wrong channel.