Update test PipelineVsFs_MultiTableDescSet
Update test and temporarily disable it.

An upcoming merge from upstream significantly changes the generated
code which is being matched in the tests. In similar situations we
usually extend tests to accept both versions, but this one is tricky,
due to the chains of instructions being changed.

The updated test can be re-enabled after the LLVM code is updated.