Optimize PatchInitializeWorkgroupMemory pass
The original PatchInitializeWorkgroupMemory pass does the initialization
by writing the LDS region occupied by a workgroup variable with multiple
null scalars/vectors in each thread. This way is ineffective if the workgroup
variable involving large number of scalars/vectors. The optmization is
to do the initialization in multi-thread mode.
1.Declare a new LDS variable (i32 array) whose size is total size of all
workgroup variables.
2.Map all workgroup variable to the new LDS variable.
3.Each thread write a null i32 value to the specified LDS offset.
4.Add an option for internal use to force the initialize

Fixes: dEQP-VK.compute.zero_initialize_workgroup_memory.*

Issue: #1302