Revert "Do not generate dummy exports for PS on GFX10"
There is a subtle interaction with fragment shaders which write to
color buffers with no attachment.  LLPC removes these outputs,
but it appears some shaders are expecting an output to happen
and relying on the results of a discard through this.

This was not picked up by Vulkan CTS testing, persumably as the
pipelines it uses are well-defined, but this behaviour does
occur in game titles.

The old dummy export behaviour was masking this problem by always
exporting when discards occur.  Revert to this behaviour while
we investigate further.

This reverts commit 533147bdb3bfd3a0dee2ec10d4c2280538e95d7d.